Got the listener largely working.
I don't have caching of price data in place, which is a bit of a shame, but I
don't really know how to handle the nested asynch of that.

Also, I don't have a good way to split functionality up by file, which sucks
but the way closures work, I need variable context passed in through param
lists that I don't own.